fixed the navigation buttons
