+++ components/About.tsx
@@ -1,0 +1,176 @@
+'use client'
+
+import { motion } from 'framer-motion'
+import { useInView } from 'framer-motion'
+import { useRef } from 'react'
+import { Users, Award, Target, Heart } from 'lucide-react'
+
+export default function About() {
+  const ref = useRef(null)
+  const isInView = useInView(ref, { once: true })
+
+  const stats = [
+    { icon: Users, label: 'Happy Clients', value: '500+' },
+    { icon: Award, label: 'Projects Completed', value: '1000+' },
+    { icon: Target, label: 'Success Rate', value: '99%' },
+    { icon: Heart, label: 'Years Experience', value: '5+' },
+  ]
+
+  const team = [
+    {
+      name: 'Mike Chen',
+      role: 'Lead Engineer',
+      image: 'https://images.unsplash.com/photo-1472099645785-5658abf4ff4e?w=300&h=300&fit=crop&crop=face',
+      bio: 'Expert in mechanical engineering and 3D printing technologies with a passion for innovation.',
+    },
+    {
+      name: 'Emily Rodriguez',
+      role: 'Design Director',
+      image: 'https://images.unsplash.com/photo-1438761681033-6461ffad8d80?w=300&h=300&fit=crop&crop=face',
+      bio: 'Creative visionary specializing in product design and user experience optimization.',
+    },
+  ]
+
+  return (
+    <>
+      <section id='about' className="section-padding bg-gradient-to-br from-primary-50 to-secondary-50">
+        <div className="container-custom">
+          <motion.div
+            initial={{ opacity: 0, y: 30 }}
+            animate={{ opacity: 1, y: 0 }}
+            transition={{ duration: 0.6 }}
+            className="text-center max-w-4xl mx-auto"
+          >
+            <h1 className="text-5xl md:text-6xl font-bold text-gray-900 mb-6">
+              About <span className="gradient-text">Proto2Plastic</span>
+            </h1>
+            <p className="text-xl text-gray-600 leading-relaxed">
+              We&apos;re a passionate team of innovators, engineers, and designers dedicated to 
+              transforming ideas into reality through cutting-edge 3D printing technology.
+            </p>
+          </motion.div>
+        </div>
+      </section>
+
+      {/* Stats Section */}
+      {/* <section className="section-padding bg-white">
+        <div className="container-custom">
+          <div className="grid grid-cols-2 md:grid-cols-4 gap-8">
+            {stats.map((stat, index) => (
+              <motion.div
+                key={stat.label}
+                initial={{ opacity: 0, y: 20 }}
+                animate={{ opacity: 1, y: 0 }}
+                transition={{ duration: 0.6, delay: index * 0.1 }}
+                className="text-center"
+              >
+                <div className="w-16 h-16 bg-gradient-primary rounded-full flex items-center justify-center mx-auto mb-4">
+                  <stat.icon className="w-8 h-8 text-white" />
+                </div>
+                <div className="text-3xl font-bold text-gray-900 mb-2">{stat.value}</div>
+                <div className="text-gray-600">{stat.label}</div>
+              </motion.div>
+            ))}
+          </div>
+        </div>
+      </section> */}
+
+      {/* Mission Section */}
+      <section className="section-padding bg-gray-50">
+        <div className="container-custom">
+          <div className="grid grid-cols-1 lg:grid-cols-2 gap-12 items-center">
+            <motion.div
+              initial={{ opacity: 0, x: -30 }}
+              animate={{ opacity: 1, x: 0 }}
+              transition={{ duration: 0.6 }}
+            >
+              <h2 className="text-4xl font-bold text-gray-900 mb-6">
+                Our Mission
+              </h2>
+              <p className="text-lg text-gray-600 mb-6 leading-relaxed">
+                To democratize access to high-quality 3D printing services, making it possible 
+                for anyone to bring their innovative ideas to life. We believe in the power of 
+                additive manufacturing to solve complex problems and create beautiful, functional objects.
+              </p>
+              <p className="text-lg text-gray-600 leading-relaxed">
+                From rapid prototyping to production-scale manufacturing, we&apos;re committed to 
+                delivering exceptional results with every project, no matter how big or small.
+              </p>
+            </motion.div>
+            <motion.div
+              initial={{ opacity: 0, x: 30 }}
+              animate={{ opacity: 1, x: 0 }}
+              transition={{ duration: 0.6, delay: 0.2 }}
+              className="bg-gradient-primary rounded-2xl p-8 text-white"
+            >
+              <h3 className="text-2xl font-bold mb-4">Why Choose Us?</h3>
+              <ul className="space-y-3">
+                <li className="flex items-center">
+                  <div className="w-2 h-2 bg-white rounded-full mr-3"></div>
+                  State-of-the-art 3D printing technology
+                </li>
+                <li className="flex items-center">
+                  <div className="w-2 h-2 bg-white rounded-full mr-3"></div>
+                  Expert team with years of experience
+                </li>
+                <li className="flex items-center">
+                  <div className="w-2 h-2 bg-white rounded-full mr-3"></div>
+                  Fast turnaround times
+                </li>
+                <li className="flex items-center">
+                  <div className="w-2 h-2 bg-white rounded-full mr-3"></div>
+                  Competitive pricing
+                </li>
+                <li className="flex items-center">
+                  <div className="w-2 h-2 bg-white rounded-full mr-3"></div>
+                  Dedicated customer support
+                </li>
+              </ul>
+            </motion.div>
+          </div>
+        </div>
+      </section>
+
+      {/* Team Section */}
+      {/* <section className="section-padding bg-white">
+        <div className="container-custom">
+          <motion.div
+            ref={ref}
+            initial={{ opacity: 0, y: 30 }}
+            animate={isInView ? { opacity: 1, y: 0 } : { opacity: 0, y: 30 }}
+            transition={{ duration: 0.6 }}
+            className="text-center mb-16"
+          >
+            <h2 className="text-4xl font-bold text-gray-900 mb-4">Meet Our Team</h2>
+            <p className="text-xl text-gray-600 max-w-3xl mx-auto">
+              The talented individuals behind our success, dedicated to bringing your ideas to life.
+            </p>
+          </motion.div>
+
+          <div className="grid grid-cols-1 md:grid-cols-3 gap-8">
+            {team.map((member, index) => (
+              <motion.div
+                key={member.name}
+                initial={{ opacity: 0, y: 30 }}
+                animate={isInView ? { opacity: 1, y: 0 } : { opacity: 0, y: 30 }}
+                transition={{ duration: 0.6, delay: index * 0.1 }}
+                className="text-center"
+              >
+                <div className="w-48 h-48 mx-auto mb-6 rounded-full overflow-hidden">
+                  <img
+                    src={member.image}
+                    alt={member.name}
+                    className="w-full h-full object-cover"
+                  />
+                </div>
+                <h3 className="text-xl font-bold text-gray-900 mb-2">{member.name}</h3>
+                <p className="text-primary-600 font-medium mb-4">{member.role}</p>
+                <p className="text-gray-600 leading-relaxed">{member.bio}</p>
+              </motion.div>
+            ))}
+          </div>
+        </div>
+      </section> */}
+  </>
+  )
+}
